refactored codebase and hardcoded a private network id to prevent chatter with ipfs mainnet
