+++ util.go
@@ -1,0 +1,163 @@
+package main
+
+import (
+	"bufio"
+	"context"
+	"crypto/rand"
+	"encoding/hex"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"strings"
+
+	"github.com/kirsle/configdir"
+	"github.com/libp2p/go-libp2p-core/peer"
+	crypto "github.com/libp2p/go-libp2p-crypto"
+	"github.com/multiformats/go-multiaddr"
+)
+
+const bootstrapPeer = "/ip4/35.224.203.143/tcp/4001/p2p/QmeRw9ZbkupTq89mrsXFX87pxzYpXR9Bmems25LPKvPbwQ"
+const ClusterKey = "2ac032e2f8161984ce91fd17e0e103e302b6b70ca3431cab5b3e8528255f330d"
+
+// ConfigSetup sets up the configuration directory
+func ConfigSetup() {
+	// Ensure config directory exists
+	configPath := configdir.LocalConfig("chatp2p")
+	er := configdir.MakePath(configPath) // Ensure it exists.
+	if er != nil {
+		panic(er)
+	}
+
+	// set up logging
+	logfile := configdir.LocalConfig("chatp2p", "chatp2p.log")
+	file, erro := os.OpenFile(logfile, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0666)
+	if erro == nil {
+		log.Out = file
+	} else {
+		log.Info("Failed to log to file, using default stderr")
+	}
+
+	fmt.Println(configPath)
+	keyfile := configdir.LocalConfig("chatp2p", ".key")
+	if _, err := os.Stat(keyfile); os.IsNotExist(err) {
+		createKey()
+	}
+	bootstrapFile := configdir.LocalConfig("chatp2p", "bootstrappers")
+	if _, err := os.Stat(bootstrapFile); os.IsNotExist(err) {
+		d1 := []byte(bootstrapPeer + "\n")
+		err := ioutil.WriteFile(bootstrapFile, d1, 0644)
+		if err != nil {
+			log.Error(err)
+			panic(err)
+		}
+	}
+}
+
+// GetKey gets the ECDSA private key from the config
+// directory
+func GetKey() crypto.PrivKey {
+	keyfile := configdir.LocalConfig("chatp2p", ".key")
+
+	// open private key file
+	content, err := ioutil.ReadFile(keyfile)
+	if err != nil {
+		panic(err)
+	}
+
+	hexString := strings.TrimSuffix(string(content), "\n")
+	decoded, err := hex.DecodeString(hexString)
+	if err != nil {
+		panic(err)
+	}
+
+	privNew, err := crypto.UnmarshalPrivateKey(decoded)
+	if err != nil {
+		panic(err)
+	}
+	return privNew
+}
+
+// CreateKey generates a new ECDSA Key Pair and stores it
+// in the config directory
+func createKey() {
+	keyfile := configdir.LocalConfig("chatp2p", ".key")
+	// Create a new ECDSA key pair for this host.
+	prvKey, _, err := crypto.GenerateECDSAKeyPair(rand.Reader)
+	if err != nil {
+		panic(err)
+	}
+	privB, err := prvKey.Bytes()
+	if err != nil {
+		panic(err)
+	}
+	pk := fmt.Sprintf("%x", string(privB))
+	ioutil.WriteFile(keyfile, []byte(pk), 0644)
+	fmt.Println("🔑 ECDSA key generated")
+}
+
+// ClusterSecret parses the hex-encoded secret string, checks that it is exactly
+// 32 bytes long and returns its value as a byte-slice.x
+func ClusterSecret() ([]byte, error) {
+	secret, err := hex.DecodeString(ClusterKey)
+	if err != nil {
+		return nil, err
+	}
+	switch secretLen := len(secret); secretLen {
+	case 0:
+		log.Warning("Cluster secret is empty, cluster will start on unprotected network.")
+		return nil, nil
+	case 32:
+		return secret, nil
+	default:
+		return nil, fmt.Errorf("input secret is %d bytes, cluster secret should be 32", secretLen)
+	}
+}
+
+// CollectBootstrapAddrInfos converts bootstrap address in config directory
+// to a slice of []peer.AddrInfo
+func CollectBootstrapAddrInfos(ctx context.Context) ([]peer.AddrInfo, error) {
+	bootstrappers := make([]string, 0)
+	bootstrapFile := configdir.LocalConfig("chatp2p", "bootstrappers")
+	file, err := os.Open(bootstrapFile)
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		bootstrappers = append(bootstrappers, scanner.Text())
+	}
+
+	if err := scanner.Err(); err != nil {
+		log.Fatal(err)
+	}
+
+	if len(bootstrappers) == 0 {
+		LogInfo("🔔 No bootstrappers defined for this node.")
+	}
+
+	addrInfoSlice := make([]peer.AddrInfo, len(bootstrappers))
+	for i, s := range bootstrappers {
+		targetAddr, err := multiaddr.NewMultiaddr(s)
+		if err != nil {
+			log.Error(err)
+			return nil, err
+		}
+
+		targetInfo, err := peer.AddrInfoFromP2pAddr(targetAddr)
+		if err != nil {
+			log.Error(err)
+			return nil, err
+		}
+		addrInfoSlice[i] = *targetInfo
+	}
+
+	return addrInfoSlice, nil
+}
+
+// LogInfo logs to console and logger
+func LogInfo(m string, args ...interface{}) {
+	fmt.Println(m, args)
+	log.Info(m, args)
+}
